requirements.txt
# 桌面灵宠项目依赖

# GUI框架
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.0
PyQt5-sip>=12.8.0

# 多媒体支持（音效）[v0.3.0]
PyQt5-Multimedia>=5.15.0

# 图像处理
Pillow>=10.0.0

# AI支持 [v0.4.0]
openai>=1.0.0            # OpenAI API
requests>=2.31.0         # HTTP 请求

# 数据库 (Python内置，无需安装)
# sqlite3

# 配置文件管理 (Python内置)
# configparser

# 日期时间处理 (Python内置)
# datetime

# 可选依赖
# ------------------

# 如果需要更高级的提醒功能
# plyer>=2.0.0  # 跨平台通知

# 如果需要音效处理
# pygame>=2.0.0  # 音频播放

# 如果需要系统集成
# pywin32>=300  # Windows系统API (仅Windows)

# 如果需要AI对话功能（参考Pet-GPT）
# openai>=0.27.0  # OpenAI API
# requests>=2.28.0  # HTTP请求

# 开发工具
# ------------------
# pytest>=7.0.0  # 测试框架
# black>=22.0.0  # 代码格式化
# flake8>=4.0.0  # 代码检查
# mypy>=0.950  # 类型检查

